Backend/requirements.txt
Flask
flask-cors
numpy
pandas
scikit-learn
tensorflow
Pillow
joblib
uvicorn
fastapi
requests